+++ lib/openhab/core_ext/java/month_day.rb
@@ -10,6 +10,7 @@ module OpenHAB
       # Extensions to MonthDay
       class MonthDay
         include Between
+        include Ephemeris
 
         class << self
           #